backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-JWT-Extended
PyMySQL
python-dotenv
Werkzeug
SQLAlchemy
    